+++ composer-show-template/src/main/java/io/piano/android/composer/showtemplate/ComposerJs.kt
@@ -4,23 +4,22 @@ import android.view.View
 import android.webkit.JavascriptInterface
 import android.webkit.WebView
 import androidx.annotation.UiThread
-import androidx.fragment.app.DialogFragment
 import io.piano.android.composer.Composer
+import io.piano.android.showhelper.BaseJsInterface
+import io.piano.android.showhelper.BaseShowController.Companion.executeJavascriptCode
+import timber.log.Timber
 
-open class ComposerJs : ClosableJs {
-    private var fragment: DialogFragment? = null
-    private var webView: WebView? = null
+open class ComposerJs : BaseJsInterface() {
     private var trackId: String = ""
 
-    internal fun init(dialogFragment: DialogFragment?, webView: WebView?, trackingId: String) {
-        fragment = dialogFragment
-        this.webView = webView
+    internal fun init(dialogFragment: ShowTemplateDialogFragment?, webView: WebView?, trackingId: String) {
+        super.init(dialogFragment, webView)
         trackId = trackingId
     }
 
     @JavascriptInterface
     @UiThread
-    open fun close(eventData: String) {
+    open fun close(eventData: String?) {
         Composer.getInstance().trackExternalEvent(trackId)
         closeOverridden(eventData)
     }
@@ -48,11 +47,19 @@ open class ComposerJs : ClosableJs {
     }
 
     @UiThread
-    override fun closeOverridden(eventData: String?) {
+    fun closeOverridden(eventData: String? = null) {
         fragment?.dismissAllowingStateLoss()
             ?: webView?.apply {
                 visibility = View.GONE
                 loadUrl("about:blank")
             }
+    }
+
+    internal fun executeJavascriptCode(code: String) {
+        val view = (fragment as? ShowTemplateDialogFragment)?.getWebView() ?: webView
+        view?.postDelayed(
+            { view.executeJavascriptCode(code) },
+            300
+        ) ?: Timber.w("We got null for webview")
     }
 }