+++ composer/src/main/java/io/piano/android/composer/model/events/EventType.kt
@@ -3,24 +3,31 @@ package io.piano.android.composer.model.events
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import io.piano.android.composer.model.DelayBy
+import io.piano.android.composer.model.DisplayMode
 import io.piano.android.composer.model.User
 
 sealed class EventType
 
+interface BaseShowType {
+    val containerSelector: String?
+    val displayMode: DisplayMode
+    val showCloseButton: Boolean
+}
+
 @JsonClass(generateAdapter = true)
 data class ExperienceExecute(
-    @JvmField val user: User?
+    val user: User?
 ) : EventType()
 
 @JsonClass(generateAdapter = true)
 data class Meter(
-    @JvmField val meterName: String,
-    @JvmField val views: Int,
-    @JvmField val viewsLeft: Int,
-    @JvmField val maxViews: Int,
-    @JvmField val totalViews: Int,
-    @JvmField val incremented: Boolean,
-    @JvmField val state: MeterState = MeterState.ACTIVE
+    val meterName: String,
+    val views: Int,
+    val viewsLeft: Int,
+    val maxViews: Int,
+    val totalViews: Int,
+    val incremented: Boolean,
+    val state: MeterState = MeterState.ACTIVE
 ) : EventType() {
     @JsonClass(generateAdapter = false)
     enum class MeterState {
@@ -33,38 +40,36 @@ object NonSite : EventType()
 
 @JsonClass(generateAdapter = true)
 data class SetResponseVariable(
-    @JvmField val responseVariables: Map<String, Any>
+    val responseVariables: Map<String, Any>
 ) : EventType()
 
 @JsonClass(generateAdapter = true)
 data class ShowLogin(
-    @JvmField val userProvider: String
+    val userProvider: String
 ) : EventType()
 
 @JsonClass(generateAdapter = true)
+data class ShowRecommendations(
+    val widgetId: String,
+    val placeholder: String?,
+    override val containerSelector: String,
+    override val displayMode: DisplayMode,
+    override val showCloseButton: Boolean,
+    val siteId: String,
+    val type: String
+) : EventType(), BaseShowType
+
+@JsonClass(generateAdapter = true)
 data class ShowTemplate(
-    @JvmField val templateId: String,
-    @JvmField val templateVariantId: String?,
-    @JvmField val displayMode: DisplayMode,
-    @JvmField val containerSelector: String?,
-    @JvmField val delayBy: DelayBy,
-    @JvmField val showCloseButton: Boolean,
-    @JvmField val url: String? = null
-) : EventType() {
-    @JsonClass(generateAdapter = false)
-    enum class DisplayMode(val mode: String) {
-        @Json(name = "modal")
-        MODAL("modal"),
-
-        @Json(name = "inline")
-        INLINE("inline");
-
-        override fun toString(): String {
-            return mode
-        }
-    }
-}
+    val templateId: String,
+    val templateVariantId: String?,
+    override val displayMode: DisplayMode,
+    override val containerSelector: String?,
+    val delayBy: DelayBy,
+    override val showCloseButton: Boolean,
+    val url: String? = null
+) : EventType(), BaseShowType
 
 data class UserSegment(
-    @JvmField val state: Boolean
+    val state: Boolean
 ) : EventType()