+++ plugins/src/main/kotlin/io/piano/android/dependencies/Deps.kt
@@ -1,7 +1,7 @@
 package io.piano.android.dependencies
 
 private object Versions {
-    const val kotlin = "1.5.31"
+    const val kotlin = "1.6.10"
 
     // AndroidX libraries
     const val compatLibrary = "1.4.1"
@@ -9,15 +9,16 @@ private object Versions {
     const val annotationsLibrary = "1.3.0"
     const val fragmentLibrary = "1.4.1"
     const val materialLibrary = "1.5.0"
-    const val lifecycleLibrary = "2.4.0"
-    const val prefsLibrary = "1.1.1"
+    const val lifecycleLibrary = "2.4.1"
+    const val prefsLibrary = "1.2.0"
+    const val multidex = "2.0.1"
 
     // Third party Libraries
-    const val facebookLogin = "12.3.0"
+    const val facebookLogin = "13.1.0"
     const val googlePlayServices = "20.1.0"
     const val retrofit = "2.6.4"
     const val okhttp = "3.12.13"
-    const val moshi = "1.12.0"
+    const val moshi = "1.13.0"
     const val timber = "5.0.1"
     const val cxense = "2.2.0"
 
@@ -27,7 +28,7 @@ private object Versions {
     const val androidxTestExtJunit = "1.1.3"
     const val mockitoKotlin = "2.2.0"
     const val mockitoCore = "3.12.1"
-    const val robolectric = "4.6.1"
+    const val robolectric = "4.7.3"
 }
 
 object Libs {
@@ -37,6 +38,7 @@ object Libs {
     const val fragments = "androidx.fragment:fragment:${Versions.fragmentLibrary}"
     const val material = "com.google.android.material:material:${Versions.materialLibrary}"
     const val lifecycleCommon = "androidx.lifecycle:lifecycle-common-java8:${Versions.lifecycleLibrary}"
+    const val multidex = "androidx.multidex:multidex:${Versions.multidex}"
 
     const val facebookLogin = "com.facebook.android:facebook-login:${Versions.facebookLogin}"
     const val googleAuth = "com.google.android.gms:play-services-auth:${Versions.googlePlayServices}"