+++ show-helper/src/main/java/io/piano/android/showhelper/BaseJsInterface.kt
@@ -1,0 +1,13 @@
+package io.piano.android.showhelper
+
+import android.webkit.WebView
+
+abstract class BaseJsInterface {
+    protected var fragment: BaseShowDialogFragment? = null
+    protected var webView: WebView? = null
+
+    protected fun init(dialogFragment: BaseShowDialogFragment?, webView: WebView?) {
+        fragment = dialogFragment
+        this.webView = webView
+    }
+}